app/product/pyproject.toml
[tool.poetry]
name = "product"
version = "0.1.0"
description = "product service"
authors = ["Deepayan Mukherjee <deepayanmukherjee1312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
motor = "^3.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.0"
boto3 = "^1.28.37"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
